Moving everything into components
